Refactor dubbo-samples-group
