+++ 2-advanced/dubbo-samples-group/dubbo-samples-group-consumer/src/test/java/org/apache/dubbo/samples/group/consumer/GroupServiceIT.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.dubbo.samples.group.consumer;
+
+import org.apache.dubbo.config.annotation.DubboReference;
+import org.apache.dubbo.samples.group.GroupService;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@SpringBootTest
+@RunWith(SpringRunner.class)
+public class GroupServiceIT {
+    @DubboReference(group = "groupA")
+    private GroupService groupServiceA;
+
+    @DubboReference(group = "groupB")
+    private GroupService groupServiceB;
+
+    @Test
+    public void testServiceA() throws Exception {
+        String result = groupServiceA.sayHello("world");
+        Assert.assertTrue(result.startsWith("Hello world"));
+        Assert.assertTrue(result.contains("group A"));
+    }
+
+    @Test
+    public void testServiceB() throws Exception {
+        String result = groupServiceB.sayHello("world");
+        Assert.assertTrue(result.startsWith("Hello world"));
+        Assert.assertTrue(result.contains("group B"));
+    }
+}
